+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update response headers to OWASP recommended values. CNJR-2116
 - Audit events to role members and graph endpoints. CONJSE-834
 - Audit events to show a role endpoint. CONJSE-833
+- Add group membership depth validation to create group membership endpoint. CNJR-13067
 
 ### Changed
 - Previously, environment variables took precedence over policy values for